+++ src/main/java/mekanism/client/render/tileentity/RenderDigitalMiner.java
@@ -19,7 +19,7 @@ public class RenderDigitalMiner extends TileEntitySpecialRenderer<TileEntityDigi
     @Override
     public void render(TileEntityDigitalMiner tileEntity, double x, double y, double z, float partialTick, int destroyStage, float alpha) {
         GlStateManager.pushMatrix();
-        GlStateManager.translate(x + 0.5, y + 1.5, z + 0.5);
+        GlStateManager.translate((float) x + 0.5F, (float) y + 1.5F, (float) z + 0.5F);
         bindTexture(MekanismUtils.getResource(ResourceType.RENDER, "DigitalMiner.png"));
 
         GLSMHelper.rotate(tileEntity.facing, 0, 180, 90, 270);
