+++ src/main/java/mekanism/generators/client/render/RenderSolarGenerator.java
@@ -14,7 +14,7 @@ public class RenderSolarGenerator extends TileEntitySpecialRenderer<TileEntitySo
     @Override
     public void render(TileEntitySolarGenerator tileEntity, double x, double y, double z, float partialTick, int destroyStage, float alpha) {
         GlStateManager.pushMatrix();
-        GlStateManager.translate(x + 0.5, y + 1.5, z + 0.5);
+        GlStateManager.translate((float) x + 0.5F, (float) y + 1.5F, (float) z + 0.5F);
         bindTexture(MekanismUtils.getResource(ResourceType.RENDER, "SolarGenerator.png"));
         GlStateManager.rotate(180, 0, 0, 1);
         model.render(0.0625F);
